MNT: migrate item classes to item.py
